StyleCop Settings.  Issue #8. * Settings.StyleCop with embedded dictionary and the 3 disabled rules per github issue 8. * Ensured we only have the one Settings.StyleCop file, at the root level, to ensure consistency throughout the codebase. * Moved TestScriptClient to its own file, and enforced remaining style rules. These were just adding explicit 'private' to the   only-implied private members, and adding XML documentation headers to all publics.
