+++ src/TestHarness/Program.cs
@@ -13,23 +13,15 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Net;
-using System.Net.Sockets;
-using System.Text;
-using System.Threading;
 using TestHarness.Commands;
 using WheelMUD.Main;
 
 namespace TestHarness
 {
-    /// <summary>
-    /// The test harness program; runs the MUD as a console application.
-    /// </summary>
+    /// <summary>The test harness program; runs the MUD as a console application.</summary>
     public class Program
     {
-        /// <summary>
-        /// Main entry point into the test harness
-        /// </summary>
+        /// <summary>Main entry point into the test harness.</summary>
         public static void Main()
         {
             string logFileName = "Log_" + DateTime.Now.ToShortDateString() + ".txt";
@@ -44,6 +36,7 @@ namespace TestHarness
 
             app.Start();
 
+            // TODO: Consider reflecting implementers of ITestHarnessCommand to keep this automatically up to date?
             ITestHarnessCommand[] commandObjects = { new HelpCommand(), new UpdateActionsCommand(), new RunTestsCommand() };
             IDictionary<string, ITestHarnessCommand> commands = new ConcurrentDictionary<string, ITestHarnessCommand>();
 
@@ -78,7 +71,6 @@ namespace TestHarness
                     {
                         display.Notify(string.Format("> Command Not Recognized. [{0}]", string.Join(" ", words)));
                     }
-
                 }
 
                 // This is for Mono compatability.
@@ -91,111 +83,11 @@ namespace TestHarness
             Console.ReadLine();
         }
 
-        /// <summary>
-        /// Notifies the user of a message.
-        /// </summary>
+        /// <summary>Notifies the user of a message.</summary>
         /// <param name="message">The message to be notified.</param>
         public void Notify(string message)
         {
             Console.WriteLine(message);
-        }
-
-        public class TestScriptClient
-        {
-            TcpClient _client;
-            NetworkStream _netstr;
-            MultiUpdater _display;
-
-            public bool Connected
-            {
-                get
-                {
-                    if (_client == null)
-                    {
-                        return false;
-                    }
-
-                    return _client.Connected;
-                }
-            }
-
-            public bool Connect(MultiUpdater display)
-            {
-                try
-                {
-                    _client = new TcpClient();
-
-                    // Use default mud port.
-                    _client.Connect(new IPEndPoint(IPAddress.Loopback, 4000));
-
-                    int attempts = 0;
-                    while (!_client.Connected && attempts++ < 10)
-                    {
-                        display.Notify("> Connecting to mud server on localhost port 4000..");
-                        Thread.Sleep(1000);
-                    }
-
-                    _display = display;
-                    _netstr = _client.GetStream();
-
-                    return true;
-                }
-                catch (Exception ex)
-                {
-                    _display.Notify("> Fatal Error: " + ex);
-                    _client = null;
-                    return false;
-                }
-            }
-
-            public bool Send(string data)
-            {
-                byte[] buf = Encoding.UTF8.GetBytes(string.Format("{0}{1}", data, Environment.NewLine));
-
-                try
-                {
-                    _netstr.Write(buf, 0, buf.Length);
-                }
-                catch (Exception ex)
-                {
-                    _display.Notify(">> ERROR: " + ex);
-                    return false;
-                }
-
-                return true;
-            }
-
-            public bool Recieve(out string data)
-            {
-                data = null;
-
-                try
-                {
-                    var buf = new byte[1024];
-
-                    _netstr.Read(buf, 0, buf.Length);
-
-                    data = Encoding.ASCII.GetString(buf);
-                }
-                catch (Exception ex)
-                {
-                    _display.Notify(">> FATAL Error: " + ex);
-                    return false;
-                }
-
-                return true;
-            }
-
-            public void Disconnect()
-            {
-                if (Connected)
-                {
-                    Send("quit");
-                    _netstr.Close();
-                    _netstr = null;
-                    _client = null;
-                }
-            }
         }
     }
 }
