+++ src/TestHarness/TestHarness.csproj
@@ -89,6 +89,7 @@
     <Compile Include="MultiUpdater.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestScriptClient.cs" />
     <Compile Include="TextLogUpdater.cs" />
   </ItemGroup>
   <ItemGroup>
